backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-api-python-client
python-dotenv
